+++ app/Http/Controllers/Backend/SubCategoryController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Backend;
 
 use App\Http\Controllers\Controller;
 use App\Models\Category;
+use App\Models\ChildCategory;
 use App\Models\SubCategory;
 use Illuminate\Http\Request;
 use App\DataTables\SubCategoryDataTable;
@@ -101,6 +102,11 @@ class SubCategoryController extends Controller
     public function destroy(string $id)
     {
         $subCategory = SubCategory::findOrFail($id);
+        $childCategory = ChildCategory::where('sub_category_id', $subCategory->id)->count();
+
+        if($childCategory > 0){
+            return response(['status' => 'error', 'message' => 'This items contain, sub items for delete this you have to delete the sub items first!']);
+        }
         $subCategory->delete();
 
         return response(['status' => 'success', 'message' => 'Deleted Successfully!']);
